Switch to the new implementation of the mock object name creation
Most mock objects now don't have the number suffix now.
Only the objects which share the same class names in the same test run will have the number suffix.

It's done to improve readability of the generated code.
In particular, it makes it easier to spot and follow the case when two or more objects of the same class
 are referenced in the same test.

 This also makes the generated code more stable and easier to review when changing the production code.